When navigating to a definition, if the cursor is potentially on the right word boundary, move the cursor left one column
