+++ integ/issues/issue20.f90
@@ -1,0 +1,9 @@
+program expr2
+implicit none
+
+integer :: x
+
+x = (2+3)*5
+print *, x
+
+end program
