+++ integ/tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
-    "module": "es2020",
+    "module": "nodenext",
     "target": "es2020",
     "lib": [
       "es2020"
     ],
-    "moduleResolution": "bundler",
+    "moduleResolution": "nodenext",
     "outDir": "../out/integ/spec",
     "sourceMap": true,
     "strict": true,
